Rename project from committee_admissions to admissions
