+++ .dockerignore
@@ -61,7 +61,7 @@ target/
 *.swp
 
 .env
-committee_admissions/files
+admissions/files
 main.js
 bundles/
 webpack-stats.json
