+++ .gitignore
@@ -61,7 +61,7 @@ target/
 *.swp
 
 .env
-committee_admissions/files
+admissions/files
 main.js
 bundles/
 webpack-stats.json
@@ -87,6 +87,6 @@ webpack-stats.json
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-committee_admissions/settings/local.py
+admissions/settings/local.py
 package-lock.json
 .vscode
