+++ admissions.ini
@@ -1,5 +1,5 @@
 [uwsgi]
-module = committee_admissions.wsgi:application
+module = admissions.wsgi:application
 master = true
 vacuum = true
 http-socket = :$(PORT)
@@ -9,7 +9,7 @@ enable-threads = true
 disable-logging = true
 processes = 4
 static-map = /media=/app/files/media
-static-map = /static=./committee_admissions/assets/
+static-map = /static=./admissions/assets/
 static-map = /static=./assets/
 
 max-requests = 50000
