+++ admissions/admissions/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class AdmissionsConfig(AppConfig):
     default_auto_field = "django.db.models.AutoField"
-    name = "committee_admissions.admissions"
+    name = "admissions.admissions"
