+++ admissions/admissions/serializers.py
@@ -1,7 +1,7 @@
 from django.utils import timezone
 from rest_framework import serializers
 
-from committee_admissions.admissions.models import (
+from admissions.admissions.models import (
     Admission,
     Group,
     GroupApplication,
