+++ admissions/assets/manifest.json
