+++ admissions/settings/development.py
@@ -58,9 +58,7 @@ DEBUG_TOOLBAR_PANELS = [
 
 INTERNAL_IPS = ["127.0.0.1"]
 
-AUTHENTICATION_BACKENDS = [
-    "committee_admissions.oauth.LegoOAuth2"
-] + AUTHENTICATION_BACKENDS
+AUTHENTICATION_BACKENDS = ["admissions.oauth.LegoOAuth2"] + AUTHENTICATION_BACKENDS
 
 SOCIAL_AUTH_LEGO_KEY = env("AUTH_LEGO_KEY", None)
 SOCIAL_AUTH_LEGO_SECRET = env("AUTH_LEGO_SECRET", None)
