+++ setup.cfg
@@ -6,12 +6,12 @@ force_grid_wrap=0
 use_parentheses=True
 line_length=88
 default_section=THIRDPARTY
-known_first_party=committee_admissions
+known_first_party=admissions
 known_django=django,rest_framework
 sections=FUTURE,STDLIB,DJANGO,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
 
 [flake8]
-exclude = venv/*,.tox/*,committee_admissions/settings/local.py,migrations,wsgi.py
+exclude = venv/*,.tox/*,admissions/settings/local.py,migrations,wsgi.py
 ignore = E203, E266, E501, W503, F403, F405, F401, C901
 max-line-length = 79
 max-complexity = 18
