+++ tox.ini
@@ -9,9 +9,9 @@ deps =
     flake8: -r{toxinidir}/requirements/flake8.txt
     isort: -r{toxinidir}/requirements/isort.txt
 commands =
-    isort: isort -c -rc committee_admissions
+    isort: isort -c -rc admissions
     flake8: flake8
-    black: black --check committee_admissions
+    black: black --check admissions
 
 [testenv:tests]
 deps =
@@ -24,7 +24,7 @@ passenv =
     DATABASE
     DATABASE_PORT
 commands =
-    coverage run --source=committee_admissions {toxinidir}/manage.py test committee_admissions
+    coverage run --source=admissions {toxinidir}/manage.py test admissions
 
 [testenv:coverage]
 basepython = python3.9
